refactor and fix bugs
