+++ public/assets/css/chat.css
@@ -1,5 +1,5 @@
 .li-list {
-  background-color: rgba(0, 0, 0, 0.6) !important;
+  background-color: rgba(255, 255, 255, 0.11) !important;
   max-width: 70%;
   border-radius: 20px;
   margin: 5px;
