+++ views/home.ejs
@@ -6,7 +6,7 @@
       name="viewport"
       content="width=device-width, initial-scale=1.0, shrink-to-fit=no"
     />
-    <title>Home - Quaternate</title>
+    <title>Quaternate</title>
     <meta
       name="description"
       content="An interactive interconnected collection"
@@ -25,9 +25,10 @@
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/simple-line-icons/2.4.1/css/simple-line-icons.min.css"
     />
+    <link rel="icon" type="image/png" href="/assets/img/pencil.png" />
   </head>
 
-  <body id="page-top" style="background: var(--purple)">
+  <body id="page-top" style="background: var(--purple); margin: 0">
     <a class="menu-toggle rounded" href="#"><i class="fa fa-bars"></i></a>
     <nav class="navbar navbar-light navbar-expand" id="sidebar-wrapper">
       <div class="container">
@@ -49,7 +50,7 @@
             <li class="nav-item sidebar-nav-item">
               <a class="nav-link js-scroll-trigger" href="#about">About</a>
             </li>
-            
+
             <li class="nav-item sidebar-nav-item">
               <a class="nav-link js-scroll-trigger" href="#portfolio"
                 >Our Products</a
@@ -69,7 +70,11 @@
     </nav>
     <header
       class="d-flex masthead"
-      style="background: url('/assets/img/portfolio-1.jpg')"
+      style="
+        background: url('/assets/img/portfolio-1.jpg');
+        background-repeat: no-repeat;
+        background-size: cover;
+      "
     >
       <div class="container text-center my-auto" style="padding: 15px30px">
         <h1 class="mb-1" style="color: var(--white)">Quaternate</h1>
@@ -79,13 +84,13 @@
         <a
           class="btn btn-dark btn-xl js-scroll-trigger"
           role="button"
-          href="/room/hall"
+          href="/cubicle/hall"
           style="
             font-size: 25px;
             background: rgb(5, 11, 18);
             padding: 15px 30px;
           "
-          >Cubicles</a
+          >My Cubicles</a
         >
         <div class="overlay"></div>
       </div>
@@ -127,20 +132,22 @@
                 <div class="caption-content">
                   <h2>ChatterBox</h2>
                   <p class="mb-0">
-                    The Dream place to your gossip sessions. Enjoy your never ending texts with your friends!
+                    The Dream place to your gossip sessions. Enjoy your never
+                    ending texts with your friends!
                   </p>
                 </div>
               </div>
               <img class="img-fluid" src="/assets/img/text.jpg" />
             </a>
-          </div>          
+          </div>
           <div class="col-lg-6">
             <a class="portfolio-item" href="#portfolio">
               <div class="caption">
                 <div class="caption-content">
                   <h2>Doodle</h2>
                   <p class="mb-0">
-                    Skribble out the real art inside your brain and let your friends see it all!
+                    Skribble out the real art inside your brain and let your
+                    friends see it all!
                   </p>
                 </div>
               </div>
@@ -153,7 +160,9 @@
                 <div class="caption-content">
                   <h2>DocBox</h2>
                   <p class="mb-0">
-                    Create and edit document files here, and finish all the stories you wanted to write with your friends! (Under Construction)
+                    Create and edit document files here, and finish all the
+                    stories you wanted to write with your friends! (Under
+                    Construction)
                   </p>
                 </div>
               </div>
@@ -165,20 +174,23 @@
               <div class="caption">
                 <div class="caption-content">
                   <h2>CamCall</h2>
-                  <p class="mb-0" >
-                    Call up your friends to see who changed his looks this lock-down - All with just a simple Cam Call!
+                  <p class="mb-0">
+                    Call up your friends to see who changed his looks this
+                    lock-down - All with just a simple Cam Call!
                   </p>
                 </div>
               </div>
               <img class="img-fluid" src="/assets/img/vc.jpg" />
             </a>
           </div>
-          
         </div>
       </div>
     </section>
     <section>
-      <div id="services" class="container text-center text-white bg-primary content-section">
+      <div
+        id="services"
+        class="container text-center text-white bg-primary content-section"
+      >
         <div class="content-section-heading">
           <h3 class="text-secondary mb-0">Services</h3>
           <h2 class="mb-5">What We Offer</h2>
@@ -216,12 +228,11 @@
           </div>
         </div>
       </div>
-        
     </section>
     <section
       class="callout"
       style="
-      margin: 30px;
+        margin: 30px;
         background: linear-gradient(
               90deg,
               rgba(255, 255, 255, 0.1) 0%,
@@ -251,40 +262,42 @@
         >
       </div>
     </section>
-    
+
     <section class="text-white bg-primary content-section">
       <div class="container text-center">
-        <h2 class="mb-4">The People you see here built this amazing website :D</h2>        
-      </div>      
+        <h2 class="mb-4">
+          The People you see here built this amazing website :D
+        </h2>
+      </div>
     </section>
     <section id="contact" class="map">
-      <div class="row" style="margin-top: 20px;">
-        <div class="col-lg-4 col-md-6 col-sm-8" style="text-align: center;color: aliceblue;"> 
-          <img class="ourimage"
-          src="/assets/img/joe.jfif"
-          height="400px"
-          >
-          <h2 style="text-align: center;">Jonathan Samuel J</h2>
+      <div class="row" style="margin-top: 20px; margin-right: 0">
+        <div
+          class="col-lg-4 col-md-6 col-sm-8"
+          style="text-align: center; color: aliceblue"
+        >
+          <img class="ourimage" src="/assets/img/joe.jfif" height="400px" />
+          <h2 style="text-align: center">Jonathan Samuel J</h2>
         </div>
-        <div class="col-lg-4 col-md-6 col-sm-8" style="text-align: center;color: aliceblue;"> 
-          <img class="ourimage"
-          src="/assets/img/creators.jpeg"
-          height="400px"
-          >
-          <h2 style="text-align: center;">Nauty and Joe</h2>
+        <div
+          class="col-lg-4 col-md-6 col-sm-8"
+          style="text-align: center; color: aliceblue"
+        >
+          <img
+            class="ourimage"
+            src="/assets/img/creators.jpeg"
+            height="400px"
+          />
+          <h2 style="text-align: center">Nauty and Joe</h2>
         </div>
-        <div class="col-lg-4 col-md-6 col-sm-8" style="text-align: center; color: aliceblue;"> 
-          <img class="ourimage"
-          src="/assets/img/vn.jpg"
-          height="400px"
-          
-          >
-          <h2 style="text-align: center;">Vaishnavee Nautiyal</h2>
+        <div
+          class="col-lg-4 col-md-6 col-sm-8"
+          style="text-align: center; color: aliceblue"
+        >
+          <img class="ourimage" src="/assets/img/vn.jpg" height="400px" />
+          <h2 style="text-align: center">Vaishnavee Nautiyal</h2>
         </div>
-
-        
       </div>
-     
     </section>
     <footer class="text-center footer">
       <div class="container">
