+++ views/register.ejs
@@ -3,6 +3,7 @@
   rel="stylesheet"
   id="bootstrap-css"
 />
+<link rel="icon" type="image/png" href="/assets/img/pencil.png" />
 <script src="//maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
 <script src="//cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
 <style>
