--- views/room.ejs
@@ -1,92 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0, shrink-to-fit=no"
-    />
-    <title>Home - Quaternate</title>
-    <meta
-      name="description"
-      content="An interactive interconnected collection"
-    />
-    <link rel="stylesheet" href="/assets/bootstrap/css/bootstrap.min.css" />
-    <link rel="stylesheet" href="/style.css" />
-  </head>
-  <body style="background: url('/assets/img/roombg.jpg'); ; background-repeat: repeat-y;
-  background-size: 100%; " >
-
-    <header
-    class="d-flex masthead"
-    style="background: url('/assets/img/4019.jpg'); background-size: 80%; background-position-x: center; background-repeat: no-repeat;"
-    >
-    <div class="container text-center my-auto" style="padding: 15px30px">
-      <h1 class="mb-1" style="color: var(--black); ">ROOMS</h1>
-      <h3 class="mb-5" style="color: var(--grey)">
-        Create and Join New Rooms Here &nbsp;
-      </h3>
-     
-      <div class="row" style="margin-bottom: 20px;">
-        <div class="col-md-6">
-          <p>
-           <button class="btn btn-primary" type="button" data-toggle="collapse" style='margin:16px; font-size: 25px; padding: 15px 30px;' data-target="#multiCollapseExample1" aria-expanded="false" aria-controls="multiCollapseExample1">Create A New Room</button>
-          </p>
-          <div class="collapse multi-collapse" id="multiCollapseExample1">
-            <div class="card card-body room">
-              <form method="POST" action="/room/create">
-                <input class="roomform" name="roomName" placeholder="Type new room name" />
-                <input type="password" class="roomform" name="password" placeholder="Type new password" />
-                <button type="submit" class="btn btn-primary">Create!!</button>
-              </form>
-            </div>
-          </div>
-        </div>
-        <div class="col-md-6">
-          <p>
-            <button class="btn btn-primary" type="button" data-toggle="collapse" style='margin:16px;font-size: 25px; padding: 15px 30px;' data-target="#multiCollapseExample2" aria-expanded="false" aria-controls="multiCollapseExample2">Join a Room</button>
-          </p>
-          <div class="collapse multi-collapse" id="multiCollapseExample2">
-            <div class="card card-body room">
-              <form method="POST" action="/room/join">
-                <input class="roomform" name="roomName" placeholder="Type joining room name" />
-                <input type="password" class="roomform" name="password" placeholder="Type the password" />
-                <button type="submit" class="btn btn-primary">Join!</button>
-              </form> 
-            </div>
-          </div>
-        </div>
-      </div>   
-      <div class="overlay"></div>
-    </div>
-    </header>  
-    <header
-      class="d-flex masthead container"
-    >
-
-    <div class="container"   >     
-        
-      
-      <div class="row ">
-        <%for(var i=0;i<user.rooms.length;i++){%>      
-          <div class="col-md-6 col-sm-6 col-lg-4">
-            <div class="card room" style="border-radius: 50px;">
-              <% var url= "/assets/img/i"; url+=1+i%6; url+=".png" %>
-              <img class="card-img-top roomcard"  alt="Collab Image" src = "<%=url%>" height="200px" >
-              <div class="card-body" style="text-align: center;">
-                <h2 class="card-title"><%=user.rooms[i].roomName%></h2>
-                <form action="/duo/<%=user.rooms[i]._id%>" method="GET"><button class="btn btn-primary" style='margin:16px' type="submit">Open Room</button></form>
-                <!-- <a href="#" class="btn btn-primary">Go somewhere</a> -->
-              </div>
-            </div>
-          </div>              
-        
-        <%}%>
-      </div>
-    </div>
-  </header>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.6.0/js/bootstrap.bundle.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.4.1/jquery.easing.min.js"></script>
-  </body>
-</html>
